requirements.txt
torch
torchvision
scikit-learn
pyyaml
pillow
matplotlib